+++ media/decoration.js
@@ -1,0 +1,526 @@
+class DecorationPreset {
+
+    //Chairs
+    static CHAIRS = {
+        OAK_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(0, 0),
+        },
+        WALNUT_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(48, 0),
+        },
+        BIRCH_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(96, 0),
+        },
+        MAHOGANY_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(144, 0),
+        },
+        RED_DINER_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(192, 0),
+        },
+        BLUE_DINER_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(240, 0),
+        },
+        COUNTRY_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(288, 0),
+        },
+        BREAKFAST_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(336, 0),
+        },
+        PINK_OFFICE_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(384, 0),
+        },
+        PURPLE_OFFICE_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(432, 0),
+        },
+        DARK_THRONE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(0, 32),
+        },
+        DINING_CHAIR_YELLOW: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(48, 32),
+        },
+        DINING_CHAIR_RED: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(96, 32),
+        },
+        GREEN_PLUSH_SEAT: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(144, 32),
+        },
+        PINK_PLUSH_SEAT: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(192, 32),
+        },
+        WINTER_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(240, 32),
+        },
+        GROOVY_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(288, 32),
+        },
+        CUTE_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(336, 32),
+        },
+        STUMP_SEAT: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(384, 32),
+        },
+        METAL_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(432, 32),
+        },
+        KING_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(0, 64),
+        },
+        CRYSTAL_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(48, 64),
+        },
+        TROPICAL_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(96, 64),
+        },
+        DESERT_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(544, 80),
+        },
+        JOJA_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(896, 432),
+        },
+        WIZARD_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(592, 1152),
+        },
+        JUNIMO_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(848, 80),
+        },
+        RETRO_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(672, 848),
+        },
+        PLASTIC_LAWN_CHAIR: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(640, 608),
+        },
+    }
+
+    static STOOLS = {
+        GREEN_OFFICE_STOOL: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(480, 0),
+        },
+        ORANGE_OFFICE_STOOL: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(496, 0),
+        },
+        GREEN_STOOL: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(480, 32),
+        },
+        BLUE_STOOL: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(496, 32),
+        },
+        JOJA_STOOL: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(896, 736),
+        },
+        WIZARD_STOOL: { 
+            size: new Vec2(16),
+            spriteSheetOffset: new Vec2(688, 1088),
+        },
+        JUNIMO_STOOL: { 
+            size: new Vec2(16),
+            spriteSheetOffset: new Vec2(880, 64),
+        },
+        RETRO_STOOL: { 
+            size: new Vec2(16),
+            spriteSheetOffset: new Vec2(640, 848),
+        },
+    }
+
+    static BENCHES = {
+        OAK_BENCH: { 
+            size: new Vec2(32, 32),
+            spriteSheetOffset: new Vec2(0, 96),
+        },
+        WALNUT_BENCH: { 
+            size: new Vec2(32, 32),
+            spriteSheetOffset: new Vec2(80, 96),
+        },
+        BIRCH_BENCH: { 
+            size: new Vec2(32, 32),
+            spriteSheetOffset: new Vec2(160, 96),
+        },
+        MAHOGANY_BENCH: { 
+            size: new Vec2(32, 32),
+            spriteSheetOffset: new Vec2(240, 96),
+        },
+        MODERN_BENCH: { 
+            size: new Vec2(32, 32),
+            spriteSheetOffset: new Vec2(320, 96),
+        },
+    }
+
+    //Tables
+    static TABLES = {
+        OAK_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(0, 560),
+        },
+        WALNUT_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(32, 560),
+        },
+        BIRCH_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(64, 560),
+        },
+        MAHOGANY_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(96, 560),
+        },
+        MODERN_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(192, 560),
+        },
+        WINTER_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(384, 560),
+        },
+        SUN_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(128, 560),
+        },
+        MOON_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(160, 560),
+        },
+        PUZZLE_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(352, 560),
+        },
+        CANDY_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(416, 560),
+        },
+        LUAU_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(448, 560),
+        },
+        DARK_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(480, 560),
+        },
+        DIVINER_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(288, 560),
+        },
+        PUB_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(224, 560),
+        },
+        LUXURY_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(256, 560),
+        },
+        NEOLITHIC_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(320, 560),
+        },
+        WINE_TABLE: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(672, 448),
+        },
+        SPIRITS_TABLE: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(672, 480),
+        },
+        DESERT_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(512, 80),
+        },
+        JOJA_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(896, 464),
+        },
+        WIZARD_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(640, 1152),
+        },
+        JUNIMO_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(896, 64),
+        },
+        RETRO_TABLE: { 
+            size: new Vec2(32, 48),
+            spriteSheetOffset: new Vec2(672, 880),
+        },
+    }
+    
+    static TEA_TABLES = {
+        OAK_TEA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(0, 608),
+        },
+        WALNUT_TEA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(32, 608),
+        },
+        BIRCH_TEA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(64, 608),
+        },
+        MAHOGANY_TEA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(96, 608),
+        },
+        MODERN_TEA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(128, 608),
+        },
+        JOJA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(864, 464),
+        },
+        WIZARD_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(688, 1168),
+        },
+        JUNIMO_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(864, 112),
+        },
+        RETRO_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(576, 912),
+        },
+    }
+
+    static END_TABLES = {
+        OAK_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(240, 688),
+        },
+        WALNUT_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(272, 688),
+        },
+        BIRCH_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(304, 688),
+        },
+        MAHOGANY_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(336, 688),
+        },
+        MODERN_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(368, 688),
+        },
+        WINTER_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(400, 688),
+        },
+        GRANDMOTHER_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(384, 688),
+        },
+        DESERT_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(544, 112),
+        },
+        JOJA_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(912, 720),
+        },
+        GRAY_JOJA_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(944, 720),
+        },
+        WIZARD_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(672, 1104),
+        },
+        JUNIMO_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(816, 112),
+        },
+        RETRO_END_TABLE: { 
+            size: new Vec2(16, 32),
+            spriteSheetOffset: new Vec2(544, 912),
+        },
+    }
+
+    static COFFE_TABLES = {
+        COFFE_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(320, 352),
+        },
+        STONE_SLAB: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(368, 352),
+        },
+        JOJA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(848, 672),
+        },
+        GRAY_JOJA_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(896, 672),
+        },
+        ELIXIR_TABLE: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(656, 1200),
+        },
+    }
+
+    //Rugs
+    static RUGS = {
+        BAMBOO_MAT: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(432, 864),
+            sortingOrder: 0,
+        },
+        BURLAP_RUG: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(224, 864),
+            sortingOrder: 0,
+        },
+        WOODCUT_RUG: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(272, 880),
+            sortingOrder: 0,
+        },
+        MONSTER_RUG: { 
+            size: new Vec2(32),
+            spriteSheetOffset: new Vec2(448, 800),
+            sortingOrder: 0,
+        },
+        OCEANIC_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(192, 608),
+            sortingOrder: 0,
+        },
+        RED_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(176, 720),
+            sortingOrder: 0,
+        },
+        PATCHWORK_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(256, 720),
+            sortingOrder: 0,
+        },
+        DARK_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(337, 720),
+            sortingOrder: 0,
+        },
+        RED_COTTAGE_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(288, 800),
+            sortingOrder: 0,
+        },
+        GREEN_COTTAGE_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(368, 800),
+            sortingOrder: 0,
+        },
+        MYSTIC_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(0, 832),
+            sortingOrder: 0,
+        },
+        NAUTICAL_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(144, 864),
+            sortingOrder: 0,
+        },
+        PIRATE_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(224, 944),
+            sortingOrder: 0,
+        },
+        FRUIT_SALAD_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(336, 944),
+            sortingOrder: 0,
+        },
+        BONE_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(192, 976),
+            sortingOrder: 0,
+        },
+        SNOWY_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(416, 976),
+            sortingOrder: 0,
+        },
+        LIGHT_GREEN_RUG: { 
+            size: new Vec2(48, 32),
+            spriteSheetOffset: new Vec2(385, 1232),
+            sortingOrder: 0,
+        },
+        LARGE_GREEN_RUG: { 
+            size: new Vec2(64, 48),
+            spriteSheetOffset: new Vec2(0, 1392),
+            sortingOrder: 0,
+        },
+        ICY_RUG: { 
+            size: new Vec2(64, 48),
+            spriteSheetOffset: new Vec2(96, 1392),
+            sortingOrder: 0,
+        },
+        OLD_WORLD_RUG: { 
+            size: new Vec2(64, 48),
+            spriteSheetOffset: new Vec2(160, 1392),
+            sortingOrder: 0,
+        },
+        LARGE_RED_RUG: { 
+            size: new Vec2(64, 48),
+            spriteSheetOffset: new Vec2(224, 1392),
+            sortingOrder: 0,
+        },
+    }
+
+}
+
+class Decoration extends GameObject {
+
+    //Constructor
+    constructor(config, preset) {
+        //Add sprite to config
+        config.sprite = `decoration/decoration.png`;
+
+        //Create game object
+        super({ ...config, ...preset });
+
+        //Add to decoration list
+        Game.decoration.push(this);
+    }
+
+    remove() {
+        super.remove();
+
+        //Remove from decoration list
+        Game.decoration.removeItem(this);
+    }
+
+}